Split ItemTrackerWidget out of AutoTrackerWindow
