+++ randovania/gui/ui_files/auto_tracker_window.ui
@@ -33,13 +33,7 @@
      </layout>
     </item>
     <item row="0" column="0">
-     <widget class="QGroupBox" name="inventory_group">
-      <property name="title">
-       <string>Inventory</string>
-      </property>
-      <property name="alignment">
-       <set>Qt::AlignLeading|Qt::AlignLeft|Qt::AlignTop</set>
-      </property>
+     <widget class="QWidget" name="item_tracker" native="true">
       <layout class="QGridLayout" name="inventory_layout"/>
      </widget>
     </item>
