+++ test/gui/test_auto_tracker_window.py
@@ -1,9 +1,11 @@
 from unittest.mock import MagicMock
 
 import pytest
+from PySide6 import QtWidgets
 
 from randovania.game_connection.connection_base import GameConnectionStatus, InventoryItem
 from randovania.gui.auto_tracker_window import AutoTrackerWindow
+from randovania.gui.widgets.item_tracker_widget import ItemTrackerWidget
 
 
 @pytest.fixture(name="window")
@@ -13,18 +15,6 @@ def auto_tracker_window(skip_qtbot):
     window = AutoTrackerWindow(connection, MagicMock())
     skip_qtbot.addWidget(window)
     return window
-
-
-def test_update_tracker_from_hook(window, echoes_resource_database):
-    # Setup
-    items = echoes_resource_database.item
-    inventory = {
-        items[i]: InventoryItem(i % 3, i % 3)
-        for i in range(len(items))
-    }
-
-    # Run
-    window._update_tracker_from_hook(inventory)
 
 
 @pytest.mark.parametrize("current_status", [GameConnectionStatus.Disconnected,
@@ -41,7 +31,6 @@ async def test_on_timer_update(current_status: GameConnectionStatus, correct_gam
     window = AutoTrackerWindow(game_connection, MagicMock())
     skip_qtbot.addWidget(window)
     window._update_timer = MagicMock()
-    window._update_tracker_from_hook = MagicMock()
 
     game_connection.get_current_inventory.return_value = inventory
     game_connection.current_status = current_status
@@ -49,24 +38,28 @@ async def test_on_timer_update(current_status: GameConnectionStatus, correct_gam
     if correct_game:
         window._current_tracker_game = game_connection.connector.game_enum
 
+    mock_update_state = MagicMock()
+    window.item_tracker.update_state = mock_update_state
+
     # Run
     await window._on_timer_update_raw()
 
     # Assert
     if current_status != GameConnectionStatus.Disconnected and correct_game:
-        window._update_tracker_from_hook.assert_called_once_with(inventory)
+        mock_update_state.assert_called_once_with(inventory)
     else:
-        window._update_tracker_from_hook.assert_not_called()
+        mock_update_state.assert_not_called()
     window._update_timer.start.assert_called_once_with()
 
 
 @pytest.mark.parametrize("name", ["Metroid Prime 1 - Game Art (Standard)", "Metroid Prime 2 - Game Art (Standard)"])
 def test_create_tracker(window: AutoTrackerWindow, name):
     window.create_tracker()
-    assert len(window._tracker_elements) == 0
+    assert type(window.item_tracker) is QtWidgets.QWidget
     for action, action_name in window._action_to_name.items():
         if action_name == name:
             action.setChecked(True)
 
     window.create_tracker()
-    assert len(window._tracker_elements) > 10
+    assert isinstance(window.item_tracker, ItemTrackerWidget)
+    assert len(window.item_tracker.tracker_elements) > 10
